+++ packages/fetch-router/src/response-helpers.ts
@@ -1,3 +1,8 @@
+export {
+  file,
+  type FileDigestFunction,
+  type FileResponseOptions,
+} from './lib/response-helpers/file.ts'
 export { html } from './lib/response-helpers/html.ts'
 export { json } from './lib/response-helpers/json.ts'
 export { redirect } from './lib/response-helpers/redirect.ts'