+++ packages/headers/src/lib/super-headers.ts
@@ -3,11 +3,15 @@ import { type AcceptEncodingInit, AcceptEncoding } from './accept-encoding.ts'
 import { type AcceptLanguageInit, AcceptLanguage } from './accept-language.ts'
 import { type CacheControlInit, CacheControl } from './cache-control.ts'
 import { type ContentDispositionInit, ContentDisposition } from './content-disposition.ts'
+import { type ContentRangeInit, ContentRange } from './content-range.ts'
 import { type ContentTypeInit, ContentType } from './content-type.ts'
 import { type CookieInit, Cookie } from './cookie.ts'
 import { canonicalHeaderName } from './header-names.ts'
 import { type HeaderValue } from './header-value.ts'
+import { type IfMatchInit, IfMatch } from './if-match.ts'
 import { type IfNoneMatchInit, IfNoneMatch } from './if-none-match.ts'
+import { IfRange } from './if-range.ts'
+import { type RangeInit, Range } from './range.ts'
 import { type SetCookieInit, SetCookie } from './set-cookie.ts'
 import { isIterable, quoteEtag } from './utils.ts'
 
@@ -35,6 +39,10 @@ interface SuperHeadersPropertyInit {
    */
   age?: string | number
   /**
+   * The [`Allow`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Allow) header value.
+   */
+  allow?: string | string[]
+  /**
    * The [`Cache-Control`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control) header value.
    */
   cacheControl?: string | CacheControlInit
@@ -58,6 +66,10 @@ interface SuperHeadersPropertyInit {
    * The [`Content-Length`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Length) header value.
    */
   contentLength?: string | number
+  /**
+   * The [`Content-Range`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Range) header value.
+   */
+  contentRange?: string | ContentRangeInit
   /**
    * The [`Content-Type`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type) header value.
    */
@@ -87,9 +99,17 @@ interface SuperHeadersPropertyInit {
    */
   ifModifiedSince?: string | DateInit
   /**
+   * The [`If-Match`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Match) header value.
+   */
+  ifMatch?: string | string[] | IfMatchInit
+  /**
    * The [`If-None-Match`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match) header value.
    */
   ifNoneMatch?: string | string[] | IfNoneMatchInit
+  /**
+   * The [`If-Range`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Range) header value.
+   */
+  ifRange?: string | Date
   /**
    * The [`If-Unmodified-Since`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Unmodified-Since) header value.
    */
@@ -102,6 +122,10 @@ interface SuperHeadersPropertyInit {
    * The [`Location`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Location) header value.
    */
   location?: string
+  /**
+   * The [`Range`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Range) header value.
+   */
+  range?: string | RangeInit
   /**
    * The [`Referer`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referer) header value.
    */
@@ -123,23 +147,28 @@ const AcceptEncodingKey = 'accept-encoding'
 const AcceptLanguageKey = 'accept-language'
 const AcceptRangesKey = 'accept-ranges'
 const AgeKey = 'age'
+const AllowKey = 'allow'
 const CacheControlKey = 'cache-control'
 const ConnectionKey = 'connection'
 const ContentDispositionKey = 'content-disposition'
 const ContentEncodingKey = 'content-encoding'
 const ContentLanguageKey = 'content-language'
 const ContentLengthKey = 'content-length'
+const ContentRangeKey = 'content-range'
 const ContentTypeKey = 'content-type'
 const CookieKey = 'cookie'
 const DateKey = 'date'
 const ETagKey = 'etag'
 const ExpiresKey = 'expires'
 const HostKey = 'host'
+const IfMatchKey = 'if-match'
 const IfModifiedSinceKey = 'if-modified-since'
 const IfNoneMatchKey = 'if-none-match'
+const IfRangeKey = 'if-range'
 const IfUnmodifiedSinceKey = 'if-unmodified-since'
 const LastModifiedKey = 'last-modified'
 const LocationKey = 'location'
+const RangeKey = 'range'
 const RefererKey = 'referer'
 const SetCookieKey = 'set-cookie'
 
@@ -418,6 +447,21 @@ export class SuperHeaders extends Headers {
   }
 
   /**
+   * The `Allow` header lists the HTTP methods that are supported by the resource.
+   *
+   * [MDN `Allow` Reference](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Allow)
+   *
+   * [HTTP/1.1 Specification](https://httpwg.org/specs/rfc9110.html#field.allow)
+   */
+  get allow(): string | null {
+    return this.#getStringValue(AllowKey)
+  }
+
+  set allow(value: string | string[] | undefined | null) {
+    this.#setStringValue(AllowKey, Array.isArray(value) ? value.join(', ') : value)
+  }
+
+  /**
    * The `Cache-Control` header contains directives for caching mechanisms in both requests and responses.
    *
    * [MDN `Cache-Control` Reference](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control)
@@ -512,6 +556,22 @@ export class SuperHeaders extends Headers {
 
   set contentLength(value: string | number | undefined | null) {
     this.#setNumberValue(ContentLengthKey, value)
+  }
+
+  /**
+   * The `Content-Range` header indicates where the content of a response body
+   * belongs in relation to a complete resource.
+   *
+   * [MDN `Content-Range` Reference](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Range)
+   *
+   * [HTTP/1.1 Specification](https://httpwg.org/specs/rfc9110.html#field.content-range)
+   */
+  get contentRange(): ContentRange {
+    return this.#getHeaderValue(ContentRangeKey, ContentRange)
+  }
+
+  set contentRange(value: string | ContentRangeInit | undefined | null) {
+    this.#setHeaderValue(ContentRangeKey, ContentRange, value)
   }
 
   /**
@@ -622,6 +682,21 @@ export class SuperHeaders extends Headers {
   }
 
   /**
+   * The `If-Match` header makes a request conditional on the presence of a matching ETag.
+   *
+   * [MDN `If-Match` Reference](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Match)
+   *
+   * [HTTP/1.1 Specification](https://datatracker.ietf.org/doc/html/rfc7232#section-3.1)
+   */
+  get ifMatch(): IfMatch {
+    return this.#getHeaderValue(IfMatchKey, IfMatch)
+  }
+
+  set ifMatch(value: string | string[] | IfMatchInit | undefined | null) {
+    this.#setHeaderValue(IfMatchKey, IfMatch, value)
+  }
+
+  /**
    * The `If-None-Match` header makes a request conditional on the absence of a matching ETag.
    *
    * [MDN `If-None-Match` Reference](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match)
@@ -634,6 +709,22 @@ export class SuperHeaders extends Headers {
 
   set ifNoneMatch(value: string | string[] | IfNoneMatchInit | undefined | null) {
     this.#setHeaderValue(IfNoneMatchKey, IfNoneMatch, value)
+  }
+
+  /**
+   * The `If-Range` header makes a range request conditional on the resource state.
+   * Can contain either an entity tag (ETag) or an HTTP date.
+   *
+   * [MDN `If-Range` Reference](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Range)
+   *
+   * [HTTP/1.1 Specification](https://datatracker.ietf.org/doc/html/rfc7233#section-3.2)
+   */
+  get ifRange(): IfRange {
+    return this.#getHeaderValue(IfRangeKey, IfRange)
+  }
+
+  set ifRange(value: string | Date | undefined | null) {
+    this.#setHeaderValue(IfRangeKey, IfRange, value)
   }
 
   /**
@@ -680,6 +771,21 @@ export class SuperHeaders extends Headers {
 
   set location(value: string | undefined | null) {
     this.#setStringValue(LocationKey, value)
+  }
+
+  /**
+   * The `Range` header indicates the part of a resource that the client wants to receive.
+   *
+   * [MDN `Range` Reference](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Range)
+   *
+   * [HTTP/1.1 Specification](https://httpwg.org/specs/rfc9110.html#field.range)
+   */
+  get range(): Range {
+    return this.#getHeaderValue(RangeKey, Range)
+  }
+
+  set range(value: string | RangeInit | undefined | null) {
+    this.#setHeaderValue(RangeKey, Range, value)
   }
 
   /**